requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
playwright==1.49.1
pydantic==2.10.4
httpx==0.28.1
python-dotenv==1.0.1
loguru==0.7.3
